refactor(DAO-2005): update btc-vault and fund module date imports
Point DepositWindowSection, deposit-history, fund-manager, and
fund-admin at the shared formatDateMonthFirst / formatDateFullMonthPadded
instead of the btc-vault-local formatters.
